# corporate-repos/psc-sre-automacao-controller/src/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSC SRE Automacao Controller",
    "description": "API do Controller de automacao SRE. Orquestra a execucao de automacoes nos Agents distribuidos por cluster, gerencia logs de execucao, autenticacao JWT e integracao com o OAS (Orquestrador de Automacao de Servicos).",
    "version": "3.6.6",
    "contact": {
      "name": "PSC SRE Automacao"
    }
  },
  "servers": [
    {
      "url": "https://sre-aut-controller.psc.{cluster}.bb.com.br",
      "description": "Controller por cluster",
      "variables": {
        "cluster": {
          "description": "Nome do cluster OpenShift",
          "default": "k8shmlbb111b"
        }
      }
    }
  ],
...
  "paths": {
...
    "/agent/list": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Listar Agents registrados",
        "description": "Retorna a lista de Agents cadastrados no Controller com namespace, cluster, ambiente e data de registro.",
        "operationId": "listAgents",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Numero maximo de agents retornados",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de agents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/agent/register": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Registrar Agent",
        "description": "Cadastra um novo Agent no inventario do Controller. Requer JWT de callback do Agent com scope REGISTER. Retorna 409 se o Agent ja estiver registrado.",
        "operationId": "registerAgent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "namespace",
                  "cluster",
                  "environment"
                ],
                "properties": {
                  "namespace": {
                    "type": "string",
                    "description": "Namespace do Agent",
                    "example": "psc-sre-automacao"
                  },
                  "cluster": {
                    "type": "string",
                    "description": "Cluster do Agent",
                    "example": "k8shmlbb111b"
                  },
                  "environment": {
                    "type": "string",
                    "enum": [
                      "desenv",
                      "hml",
                      "prod"
                    ],
                    "description": "Ambiente do Agent"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Agent created."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "namespace": {
                          "type": "string"
                        },
                        "cluster": {
                          "type": "string"
                        },
                        "environment": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Payload invalido - namespace, cluster ou environment ausente/invalido"
          },
          "409": {
            "description": "Agent ja registrado com este cluster/namespace/environment"
          }
        }
      }
    },
    "/agent/execute": {
      "post": {
        "tags": [
          "Execute"
        ],
        "summary": "Executar automacao no Agent (rota legada)",
        "description": "Despacha a execucao de uma funcao de automacao para o Agent do cluster/namespace informado. Suporta modo sincrono (sync) e assincrono (async). Requer autenticacao JWT com scope EXECUTE.",
        "operationId": "executeAgent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "description": "Modo de execucao: sync (aguarda resultado) ou async (retorna imediatamente com execId)",
            "schema": {
              "type": "string",
              "enum": [
                "sync",
                "async"
              ],
              "default": "async"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "cluster",
                  "namespace",
                  "function"
                ],
                "properties": {
                  "cluster": {
                    "type": "string",
                    "description": "Nome do cluster de destino",
                    "example": "k8shmlbb111b"
                  },
                  "namespace": {
                    "type": "string",
                    "description": "Namespace do workload no cluster",
                    "example": "psc-sre-automacao"
                  },
                  "function": {
                    "type": "string",
                    "description": "Nome da funcao de automacao a ser executada no Agent",
                    "example": "health_check"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Execucao concluida (modo sync)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSnapshot"
                }
              }
            }
          },
          "202": {
            "description": "Execucao aceita (modo async)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AsyncExecuteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Campos obrigatorios ausentes ou invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent nao registrado para o cluster/namespace informado"
          },
          "502": {
            "description": "Falha na comunicacao com o Agent"
          },
          "504": {
            "description": "Timeout aguardando resposta do Agent"
          }
        }
      },
      "get": {
        "tags": [
          "Execution Logs"
        ],
        "summary": "Consultar status e logs de uma execucao",
        "description": "Retorna o snapshot de execucao com status atual, entradas de log e metadados. Suporta filtros por data e limite de entradas.",
        "operationId": "getExecutionLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "query",
            "description": "ID unico da execucao (recomendado)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "execId",
            "in": "query",
            "description": "ID da execucao (alternativa ao uuid)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "day",
            "in": "query",
            "description": "Filtrar logs por dia (formato YYYY-MM-DD)",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "description": "Filtrar logs a partir desta data/hora (ISO 8601)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "description": "Filtrar logs ate esta data/hora (ISO 8601)",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Numero maximo de entradas retornadas",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot da execucao com logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "uuid ou execId nao informado"
          }
        }
      }
    },
    "/agent/execute/logs": {
      "post": {
        "tags": [
          "Execution Logs"
        ],
        "summary": "Enviar logs de execucao (callback do Agent)",
        "description": "Recebe entradas de log enviadas pelo Agent durante a execucao de uma automacao. Requer JWT de callback do Agent com scope SEND.",
        "operationId": "pushExecutionLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "execId",
                  "entries"
                ],
                "properties": {
                  "execId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "ID da execucao"
                  },
                  "source": {
                    "type": "string",
                    "description": "Origem do log (ex: agent, controller)",
                    "default": "agent"
                  },
                  "from": {
                    "type": "string",
                    "description": "Identificador de quem envia o log"
                  },
                  "entries": {
                    "type": "array",
                    "description": "Lista de entradas de log",
                    "items": {
                      "$ref": "#/components/schemas/LogEntry"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Logs recebidos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "received": {
                      "type": "integer",
                      "example": 5
                    },
                    "execId": {
                      "type": "string"
                    },
                    "from": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "execId ou entries ausente/invalido"
          },
          "403": {
            "description": "execId do JWT nao confere com o do body"
          }
        }
      }
    },
    "/agent/errors": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Consultar log de erros",
        "description": "Retorna as ultimas linhas do arquivo local de erros/log do Controller.",
        "operationId": "getAgentErrors",
        "parameters": [
          {
            "name": "lines",
            "in": "query",
            "description": "Numero de linhas a retornar",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Linhas do log de erros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "lines": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cronjob/result": {
      "post": {
        "tags": [
          "Cronjob"
        ],
        "summary": "Receber resultado de execucao de cronjob",
        "description": "Recebe o payload final da execucao do cronjob enviado pelo Agent, adapta de compliance_status para o formato de execution log e armazena usando a infraestrutura existente de snapshot. Indexa por namespace + execId.",
        "operationId": "receiveCronjobResult",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "compliance_status",
                  "namespace",
                  "cluster_type",
                  "timestamp",
                  "execId"
                ],
                "properties": {
                  "compliance_status": {
                    "type": "string",
                    "enum": [
                      "success",
                      "failed",
                      "error"
                    ],
                    "description": "Status da execucao do cronjob"
                  },
                  "namespace": {
                    "type": "string",
                    "description": "Namespace onde o cronjob executou"
                  },
                  "cluster_type": {
                    "type": "string",
                    "description": "Tipo do cluster (ex: origem, destino)"
                  },
                  "timestamp": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Timestamp da execucao"
                  },
                  "execId": {
                    "type": "string",
                    "description": "ID unico da execucao"
                  },
                  "captured_data": {
                    "type": "object",
                    "description": "Dados capturados (quando compliance_status=success)"
                  },
                  "failures": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "Lista de falhas (quando compliance_status=failed)"
                  },
                  "errors": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    },
                    "description": "Lista de erros (quando compliance_status=error)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado armazenado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "execId": {
                      "type": "string"
                    },
                    "namespace": {
                      "type": "string"
                    },
                    "compliance_status": {
                      "type": "string"
                    },
                    "indexed": {
                      "type": "boolean"
                    },
                    "statusEndpoint": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Payload invalido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao armazenar resultado"
          }
        }
      }
    },
    "/api/cronjob/status/{execId}": {
      "get": {
        "tags": [
          "Cronjob"
        ],
        "summary": "Consultar resultado de execucao de cronjob",
        "description": "Retorna o resultado armazenado de uma execucao de cronjob pelo execId. Usa a infraestrutura existente de ExecutionSnapshot.",
        "operationId": "getCronjobStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "execId",
            "in": "path",
            "required": true,
            "description": "ID da execucao",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado da execucao",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSnapshot"
                }
              }
            }
          },
          "400": {
            "description": "execId ausente"
          },
          "500": {
            "description": "Erro interno"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT emitido por POST /auth/token. Algoritmo HS256, issuer=psc-sre-automacao-controller."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API Key para emissao de tokens JWT e consulta de scopes. Configurada via AUTH_API_KEYS_SCOPES no Kubernetes Secret."
      },
      "internalOrigin": {
        "type": "apiKey",
        "in": "header",
        "name": "x-techbb-namespace",
        "description": "Autenticacao por origem interna do Kubernetes. Quando os headers x-techbb-namespace e x-techbb-service-account correspondem ao caller confiavel configurado, a autenticacao JWT e dispensada."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Mensagem de erro"
          },
          "detail": {
            "type": "string",
            "description": "Detalhes adicionais do erro"
          },
          "execId": {
            "type": "string",
            "format": "uuid",
            "description": "ID da execucao, quando disponivel"
          }
        }
      },
...
      "AgentListResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "count": {
            "type": "integer"
          },
          "agents": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "namespace": {
                  "type": "string"
                },
                "cluster": {
                  "type": "string"
                },
                "environment": {
                  "type": "string",
                  "enum": [
                    "desenv",
                    "hml",
                    "prod"
                  ]
                },
                "registeredAt": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        }
      },
...
      "AsyncExecuteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "execId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "type": "string",
            "example": "async"
          },
          "status": {
            "type": "string",
            "example": "RUNNING"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "message": {
            "type": "string",
            "example": "Request accepted. The Agent will process this execution asynchronously."
          }
        }
      },
...
      "ExecutionSnapshot": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "execId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "DONE",
              "ERROR"
            ]
          },
          "statusLabel": {
            "type": "string",
            "example": "Running"
          },
          "finished": {
            "type": "boolean"
          },
          "lastUpdate": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer",
            "description": "Quantidade de entradas de log"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogEntry"
            }
          }
        }
      },
      "LogEntry": {
        "type": "object",
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp da entrada"
          },
          "level": {
            "type": "string",
            "description": "Nivel do log (INFO, WARN, ERROR)"
          },
          "message": {
            "type": "string",
            "description": "Mensagem do log"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "DONE",
              "ERROR"
            ],
            "description": "Status da execucao nesta entrada"
          },
          "step": {
            "type": "integer",
            "description": "Numero do passo da automacao"
          },
          "execId": {
            "type": "string",
            "description": "ID da execucao"
          },
          "source": {
            "type": "string",
            "description": "Origem do log (agent, controller, trace)"
          },
          "from": {
            "type": "string",
            "description": "Identificador de quem enviou o log"
          },
          "data": {
            "description": "Dados adicionais da entrada"
          }
        }
      }
    }
  }
}